requirements.txt
Django==4.2.7
djangorestframework==3.14.0
requests==2.31.0
google-generativeai==0.3.2
django-cors-headers==4.3.1
python-dotenv==1.0.0
geopy==2.4.0
folium==0.15.0
